src/pyproject.toml
[tool.poetry]
name = "aai_temp"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Derek Ding <derek.ding@iag.com.au>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
sqlalchemy = "1.4.46"
pandas = "1.4.3"
numpy = "1.23.1"
pytz = "2023.3"
scikit-learn = "1.3.0"
scipy = "1.12"
SQLAlchemy = "1.4.46"
python-dotenv = "0.20.0"
wget = "^3.2"

lightgbm = "3.3.2"
umap = "^0.1.1"
nltk = "3.7"
gensim = "4.2.0"
fuzzywuzzy = "0.18.0"

dbt-postgres = "1.7.2"
pyprojroot = "^0.3.0"
pyarrow = "^15.0.0"
google-cloud-aiplatform = "^1.42.1"
jsonlines = "^4.0.0"
google-cloud-secret-manager = "^2.18.2"
nbformat = "^5.9.2"
nbclient = "^0.9.0"
quarto = "^0.1.0"
lime = "^0.2.0.1"
markupsafe = "2.0.1"
protobuf = "4.25.3"
great-expectations = "^0.18.12"
numba = "0.58.1"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pyfarmhash = "^0.3.2"
ydata-profiling = "^4.6.5"
ipywidgets = "^8.1.1"
shap = "^0.45.0"
selenium = "^4.16.0"
boto3 = "1.24.95"
waterfallcharts = "^3.8"
python-Levenshtein = "^0.20.9"
wordcloud = "1.9.2"
optuna = "2.10.1"
matplotlib = "3.6.1"
seaborn = "0.11.2"
fastparquet = "0.8.3"
joblib = "1.2.0"
ipykernel = "^6.22.0"

[[tool.poetry.source]]
name = "nexus"
url = "https://nexus3.auiag.corp/repos/repository/ddo-pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[tool.poetry.group.dev.dependencies]
tabulate = "^0.9.0"
plotly = "^5.19.0"
xgboost = "^2.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
